requirements.txt
# Core dependencies
PyYAML>=6.0
boto3>=1.26.0

# Database client dependencies (recommended installation as fallback for system clients)
# MySQL support
PyMySQL>=1.0.0

# PostgreSQL support
psycopg2-binary>=2.9.0
